# src/assets/open-api/my_server.openapi.json
{"openapi":"3.0.0","paths":{"/api/auth/login":{"post":{"operationId":"AuthController_signIn","parameters":[],"responses":{"200":{"description":""}},"tags":["Auth"]}},"/api/user/create":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"创建用户","properties":{"statusCode":{"type":"number","default":"0"},"message":{"type":"string","default":"success"},"data":{"$ref":"#/components/schemas/UserModel"}}}}}}},"summary":"创建用户","tags":["用户"]}},"/api/user/test-auth":{"get":{"operationId":"UserController_testAuth","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"测试鉴权","properties":{"statusCode":{"type":"number","default":"0"},"message":{"type":"string","default":"success"},"data":{"$ref":"#/components/schemas/UserModel"}}}}}}},"summary":"测试鉴权","tags":["用户"]}},"/api/file/upload":{"post":{"operationId":"FileController_upload","parameters":[],"responses":{"201":{"description":""}},"tags":["File"]}},"/api/file/image-upload":{"post":{"operationId":"FileController_uploadImage","parameters":[],"responses":{"201":{"description":""}},"tags":["File"]}},"/api/file/image-upload-base64":{"post":{"operationId":"FileController_uploadImageBase64","parameters":[],"responses":{"201":{"description":""}},"tags":["File"]}},"/api/todo-list/list":{"get":{"operationId":"TodoListController_getList","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"返回列表","allOf":[{"$ref":"#/components/schemas/ResponseListDto"},{"properties":{"statusCode":{"type":"number","default":"0"},"message":{"type":"string","default":"success"},"data":{"type":"array","items":{"$ref":"#/components/schemas/TodoListModel"}}}}]}}}}},"tags":["待办事项"]}},"/api/todo-list/list-by-page":{"get":{"operationId":"TodoListController_getListByPage","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"分页返回列表","allOf":[{"$ref":"#/components/schemas/ResponsePaginatedDto"},{"properties":{"statusCode":{"type":"number","default":"0"},"message":{"type":"string","default":"success"},"data":{"allOf":[{"$ref":"#/components/schemas/PaginatedDto"},{"properties":{"results":{"type":"array","items":{"$ref":"#/components/schemas/TodoListModel"}}}}],"type":"object","items":{"$ref":"#/components/schemas/PaginatedDto"}}}}]}}}}},"tags":["待办事项"]}},"/api/todo-list/{id}":{"get":{"operationId":"TodoListController_getListById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"title":"返回单个","properties":{"statusCode":{"type":"number","default":"0"},"message":{"type":"string","default":"success"},"data":{"$ref":"#/components/schemas/TodoListModel"}}}}}}},"tags":["待办事项"]}}},"info":{"title":"My Server","description":"my Server API description","version":"1.0","contact":{}},"tags":[{"name":"my","description":""}],"servers":[],"components":{"schemas":{"ResponseObjDto":{"type":"object","properties":{"statusCode":{"type":"number","description":"状态码"},"message":{"type":"string","description":"消息"},"data":{"type":"object","description":"结果"}},"required":["statusCode","message","data"]},"ObjectId":{"type":"object","properties":{}},"UserModel":{"type":"object","properties":{"_id":{"description":"id","allOf":[{"$ref":"#/components/schemas/ObjectId"}]},"createdAt":{"format":"date-time","type":"string","description":"创建时间"},"updatedAt":{"format":"date-time","type":"string","description":"更新时间"},"username":{"type":"string","description":"用户名"},"nickname":{"type":"string","description":"昵称"}},"required":["_id","createdAt","updatedAt","username","nickname"]},"CreateUserDto":{"type":"object","properties":{"username":{"type":"string","description":"用户名"},"password":{"type":"string","description":"密码"},"nickname":{"type":"string","description":"昵称"}},"required":["username","password"]},"ResponseListDto":{"type":"object","properties":{"statusCode":{"type":"number","description":"状态码"},"message":{"type":"string","description":"消息"},"data":{"type":"object","description":"结果"}},"required":["statusCode","message","data"]},"TodoListModel":{"type":"object","properties":{"title":{"type":"string","description":"标题"},"done":{"type":"boolean","description":"是否完成"}},"required":["title","done"]},"ResponsePaginatedDto":{"type":"object","properties":{"statusCode":{"type":"number","description":"状态码"},"message":{"type":"string","description":"消息"},"data":{"type":"object","description":"结果"}},"required":["statusCode","message","data"]},"PaginatedDto":{"type":"object","properties":{"total":{"type":"number"},"page_size":{"type":"number"},"current":{"type":"number"},"results":{"type":"array","items":{"type":"string"}}},"required":["total","page_size","current","results"]}}}}